+++ src/main/java/com/reborn/server/domain/job/api/ForecastApi.java
@@ -17,14 +17,12 @@ import org.springframework.web.client.RestTemplate;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/job")
+@RequestMapping("/api/jobs")
 @RequiredArgsConstructor
 public class ForecastApi {
 
     private final RestTemplate restTemplate;
-
     private final JobPostService jobPostService;
-    private final JobPostDetailService jobPostDetailService;
 
     @Value("${openApi.service-key}")
     private String serviceKey;
@@ -32,20 +30,10 @@ public class ForecastApi {
     @Value("${openApi.base-url}")
     private String baseUrl;
 
-    @Value("${openApi.detail-url}")
-    private String detailUrl;
-
-
     @GetMapping
-    public ResponseEntity<List<JobPostDto>> getJobPost() throws Exception {
+    public ResponseEntity<Void> reloadJobPost() throws Exception {
         jobPostService.syncJobData(restTemplate, baseUrl, serviceKey);
         return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    @GetMapping("/detail")
-    public ResponseEntity<JobPostDetailDto> getJobPostDetail(@RequestParam String jobId) throws Exception {
-        JobPostDetailDto jobPostDetail = jobPostDetailService.getJobPostDetail(restTemplate, detailUrl, serviceKey, jobId);
-        return new ResponseEntity<>(jobPostDetail,HttpStatus.OK);
     }
 }
