+++ src/main/java/com/reborn/server/domain/user/dao/UserRepository.java
@@ -7,6 +7,5 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository <User, Long> {
     Optional<User> findByEmail(String email);
-
     Optional<User> findByName(String userName);
 }
